Updated linter to use modules imported via poetry
`pipenv` with the `--system` argument was allowing for the packages
to be installed at the system level versus in a virtualenv which
resulted in `isort` throwing an error that the command did not exist.

The workflow has been modified to invoke poetry to utilize `isort` and
`black`.

Additionally, the `setup.py` and `setup.cfg` files have been re-added
back into this branch for compatibility.