Changed output of function and bugfix
Change: `get_machine_sync_details` has been changed back to returning
a dict of values.  The machine ID will be the dict key for the CE
machines returned from a project.

Bugfix: CE machines not in a `STARTED` state may have an empty
`lastSeenDateTime` in its returned response from the API.  This is
better guarded against in the report result.